Add files via upload
This project is a movie recommendation system developed as part of the Bharat Intern program. The system leverages machine learning algorithms to suggest personalized movie recommendations based on user preferences and viewing history. It utilizes data from popular movie databases and employs various collaborative filtering techniques to enhance the accuracy and relevance of recommendations. By providing users with tailored movie suggestions, this project aims to improve the overall viewing experience and help users discover new films they are likely to enjoy. The project includes data preprocessing, model training, and evaluation, as well as a user-friendly interface for easy interaction. Whether you're a movie enthusiast or a data science student, this project offers an insightful exploration of recommendation systems in the entertainment industry.